requirements.txt
annotated-types
anyio
asgiref
boto3
botocore
certifi
colorama
distro
dj-database-url
Django==4.0.1
django-environ
django-storages
djangorestframework
exceptiongroup
gunicorn
httpcore
jmespath
openai
packaging
pillow
psycopg2-binary
pydantic_core
six
django-cors-headers==4.5.0
django-filter==24.3